requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

PyMuPDF==1.23.26
pdf2image==1.17.0
python-pptx==0.6.23
openpyxl==3.1.2
python-docx==1.1.0
Pillow==10.2.0

openai==1.12.0
langfuse==2.23.0
qdrant-client==1.7.3
redis==5.0.1
celery[redis]==5.3.6

networkx==3.2.1
numpy==1.26.3
scikit-learn==1.4.0
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
